Switch mac/hid.c to use threads.
The previous mac/hid.c implementation was kind of problematic. It worked
fine on 10.5, but on 10.7, I have gotten reports of really weird stuff
happening, such as the report callback being called with invalid context
data.

My theory is that somewhere, some other CFRunLoopRun*() call is managing
to call the report callback with some other context. Needless to say,
this is bad. Putting the run loop on a separate thread should isolate it
from any other parts of the program calling the RunLoopRun*() functions.

Further, with the previous implementation, the OS only seemed to buffer
up 4 reports in between calls to CFRunLoopRun*(), causing the potential
for reports to get lost. The new method, being on a separate thread,
will now buffer up to a configurable 30 reports.