Callback fix for OSX 10.7 Lion.
The callback's don't get removed from the callback list when you call
IOHIDDeviceClose() in 10.7 like they used to. This means that
the report callback must be unregistered at hid_close() time. Further
the unregister will not work unless you pass _exactly_ the same
parameters into IOHIDDeviceRegisterInputReportCallback() that were
passed in the first time (with the exception of the function being
NULL). Thanks Apple for your stellar documentation and commitment
to backwards compatibility.